+++ otherfile/version_check.py
@@ -6,7 +6,7 @@ import codecs
 from art.art_param import *
 
 Failed = 0
-VERSION = "5.0"
+VERSION = "5.1"
 
 README_ITEMS = ['<td align="center">{0}</td>'.format(str(FONT_COUNTER)),
                 '<img src="https://img.shields.io/badge/Art List-{0}-orange.svg">'.format(str(ART_COUNTER)),
@@ -83,7 +83,7 @@ if __name__ == "__main__":
                 break
     except Exception as e:
         Failed += 1
-        print("Error in " + file_name + "\n" + "Message : " + str(e))
+        print("Error in README.md" + "\n" + "Message : " + str(e))
     if Failed == 0:
         print_result(False)
         sys.exit(0)